Float arithmetics are actually faster! (#1)
* Float arithmetics are actually faster!

* Cleanup

* time_res should be in seconds

* Fixes

+++ 12tet.c
@@ -1,14 +1,14 @@
 #include "12tet.h"
 
 void generate_12tet_bands(FreqBand *bands, int start_midi_note, int num_notes,
-                          double width) {
+                          float width) {
   for (int i = 0; i < num_notes; i++) {
     int m = start_midi_note + i;
 
-    double ctr = 440.0 * pow(2.0, (m - 69.0) / 12.0);
+    float ctr = 440.0f * powf(2.0f, (float)(m - 69) / 12.0f);
 
     bands[i].ctr = ctr;
-    bands[i].lo = ctr * pow(2.0, -width / 24.0);
-    bands[i].hi = ctr * pow(2.0, width / 24.0);
+    bands[i].lo = ctr * powf(2.0f, -width / 24.0f);
+    bands[i].hi = ctr * powf(2.0f, width / 24.0f);
   }
 }
