+++ 12tet.h
@@ -4,6 +4,6 @@
 #include "vqsdft.h"
 
 void generate_12tet_bands(FreqBand *bands, int start_midi_note, int num_notes,
-                          double width);
+                          float width);
 
 #endif
